html/syntax/charset/inheritance-bogus-meta.html

<!doctype html>
<meta charset=windows-1253>
<title>bogus meta charset in iframe inheriting windows-1253 from parent</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<iframe src=></iframe>
<script>
["resources/bogus-charset.html",
 "resources/bogus-charset-http.py"].forEach(url => {
  async_test(t => {
    const frame = document.createElement("iframe");
    frame.src = url;
    t.add_cleanup(() => frame.remove());
    frame.onload = t.step_func_done(() => {
      // 0x00A2 in windows-1252 is at the same position as 0x0386 in windows-1253
      assert_equals(frame.contentDocument.body.textContent, "¢\n");
    });
    document.body.append(frame);
  });
});

  async_test(t => {
    const frame = document.createElement("iframe");
    frame.src = "resources/bogus-charset-http-valid-meta.py";
    t.add_cleanup(() => frame.remove());
    frame.onload = t.step_func_done(() => {
      assert_equals(frame.contentDocument.body.textContent, "\u045E\n");
    });
    document.body.append(frame);
  });
</script>
